+++ test/src/color.test.ts
@@ -1,5 +1,4 @@
 import assert from 'assert';
-import {namedColors} from '@bhsd/common';
 import {getDelimiter, alphaToString, colorToString} from '../../dist/color.js';
 import type {WidgetOptions} from '../../dist/types';
 
@@ -16,27 +15,27 @@ const colorTest = (
 	const options = {colorType, color: [0, 0, 0]} as WidgetOptions,
 		value = '#ff0000';
 	assert.strictEqual(
-		colorToString({...options, alpha: 1, legacy: false}, value, namedColors),
+		colorToString({...options, alpha: 1, legacy: false}, value),
 		expected[0],
 	);
 	assert.strictEqual(
-		colorToString({...options, alpha: 1, legacy: true, spaced: true}, value, namedColors),
+		colorToString({...options, alpha: 1, legacy: true, spaced: true}, value),
 		expected[1],
 	);
 	assert.strictEqual(
-		colorToString({...options, alpha: 1, legacy: true, spaced: false}, value, namedColors),
+		colorToString({...options, alpha: 1, legacy: true, spaced: false}, value),
 		expected[2],
 	);
 	assert.strictEqual(
-		colorToString({...options, alpha: 0.555, legacy: false}, value, namedColors),
+		colorToString({...options, alpha: 0.555, legacy: false}, value),
 		expected[3],
 	);
 	assert.strictEqual(
-		colorToString({...options, alpha: 0.555, legacy: true, spaced: true}, value, namedColors),
+		colorToString({...options, alpha: 0.555, legacy: true, spaced: true}, value),
 		expected[4],
 	);
 	assert.strictEqual(
-		colorToString({...options, alpha: 0.555, legacy: true, spaced: false}, value, namedColors),
+		colorToString({...options, alpha: 0.555, legacy: true, spaced: false}, value),
 		expected[5],
 	);
 };